MRC/M+/M-
Final Update for calculator app
