+++ calculator/app/src/main/java/com/example/calculator/models/Functions.kt
@@ -1,6 +1,5 @@
-package com.example.calculator
+package com.example.calculator.models
 
-import android.util.Log
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
@@ -15,6 +14,8 @@ class Functions : ViewModel() {
     private var previousNumber: Double? = null
     private var operator: String? = null
     private var isCalculatorOn: Boolean = false
+    private var mModeOn: Boolean = false
+    private var mNumber: Double? = null
 
     private val lastTwoButtons = mutableListOf<String>()
 
@@ -24,58 +25,72 @@ class Functions : ViewModel() {
 
         when (btn) {
             "ON/C" -> {
-                //Ligar e resetar os valores
                 _ResultText.value = "0.0"
                 currentNumber = null
                 previousNumber = null
                 operator = null
                 isCalculatorOn = true
-                Log.i("ON/C antes / atual", previousNumber.toString() + " / " + currentNumber.toString())
-                return
+                mModeOn = false
+                mNumber = null
             }
 
             "M+" -> {
-                // Verifica se a calculadora está ligada
                 if (isCalculatorOn) {
-                    val currentTextValue = _ResultText.value?.toDoubleOrNull() ?: 0.0
-                    previousNumber = (previousNumber ?: 0.0) + currentTextValue
-                    _ResultText.value = previousNumber.toString()
+                    currentNumber = _ResultText.value?.toDoubleOrNull()
+                    if (previousNumber != null && currentNumber != null && operator != null) {
+                        previousNumber = performOperation(previousNumber!!, currentNumber!!, operator!!)
+                        _ResultText.value = previousNumber.toString()
+                    } else {
+                        previousNumber = currentNumber
+                    }
+
+                    mNumber = (mNumber ?: 0.0) + (previousNumber ?: 0.0)
+                    _ResultText.value = "0.0"
+                    operator = null
                 }
-                Log.i("M+ antes / atual", previousNumber.toString() + " / " + currentNumber.toString())
                 return
             }
 
             "M-" -> {
-                // Verifica se a calculadora está ligada
                 if (isCalculatorOn) {
-                    val currentTextValue = _ResultText.value?.toDoubleOrNull() ?: 0.0
-                    previousNumber = (previousNumber ?: 0.0) - currentTextValue
-                    _ResultText.value = previousNumber.toString()
+                    currentNumber = _ResultText.value?.toDoubleOrNull()
+
+                    if (previousNumber != null && currentNumber != null && operator != null) {
+                        previousNumber = performOperation(previousNumber!!, currentNumber!!, operator!!)
+                        _ResultText.value = previousNumber.toString()
+                    } else {
+                        previousNumber = currentNumber
+                    }
+
+                    if(mNumber == null){
+                        mNumber = previousNumber
+                    } else if (previousNumber != null) {
+                        mNumber = (mNumber ?: 0.0) - previousNumber!!
+                    }
+
+
+                    _ResultText.value = "0.0"
+                    operator = null
                 }
-                Log.i("M- antes / atual", previousNumber.toString() + " / " + currentNumber.toString())
                 return
             }
+
+
 
             "MRC" -> {
 
                 if (isCalculatorOn) {
                     if (lastTwoButtons.count { it == "MRC" } == 2) {
-                        // Se os dois últimos botões foram "MRC", reseta os valores
-                        currentNumber = null
-                        previousNumber = null
+                        mNumber = null
                         _ResultText.value = "0"
                     } else {
-                        currentNumber = previousNumber
-                        previousNumber = null
-                        _ResultText.value = currentNumber?.toString() ?: "0"
+                        _ResultText.value = mNumber?.toString() ?: "0"
                     }
                 }
-                Log.i("MRC antes / atual", previousNumber.toString() + " / " + currentNumber.toString())
                 return
             }
 
             "+" -> {
-                // Só faz algo se a calculadora estiver ligada
                 if (isCalculatorOn) {
                     currentNumber = _ResultText.value?.toDoubleOrNull()
                     if (previousNumber != null && currentNumber != null && operator != null) {
@@ -85,11 +100,9 @@ class Functions : ViewModel() {
                         previousNumber = currentNumber
                     }
                     operator = btn
-                    currentNumber = null // Limpa o currentNumber para o próximo número
-                    _ResultText.value = "0.0" //Prepara para o próximo número
+                    currentNumber = null
+                    _ResultText.value = "0.0"
                 }
-                Log.i("+ antes / atual", previousNumber.toString() + " / " + currentNumber.toString())
-                return
             }
 
             "-" -> {
@@ -102,11 +115,9 @@ class Functions : ViewModel() {
                         previousNumber = currentNumber
                     }
                     operator = btn
-                    currentNumber = null // Limpa o currentNumber para o próximo número
-                    _ResultText.value = "0.0" // Prepara para o próximo número
+                    currentNumber = null
+                    _ResultText.value = "0.0"
                 }
-                Log.i("- antes / atual", previousNumber.toString() + " / " + currentNumber.toString())
-                return
             }
 
             "x" -> {
@@ -119,11 +130,9 @@ class Functions : ViewModel() {
                         previousNumber = currentNumber
                     }
                     operator = btn
-                    currentNumber = null // Limpa o currentNumber para o próximo número
-                    _ResultText.value = "0.0" // Prepara para o próximo número
+                    currentNumber = null
+                    _ResultText.value = "0.0"
                 }
-                Log.i("x antes / atual", previousNumber.toString() + " / " + currentNumber.toString())
-                return
             }
 
             "÷" -> {
@@ -136,11 +145,9 @@ class Functions : ViewModel() {
                         previousNumber = currentNumber
                     }
                     operator = btn
-                    currentNumber = null // Limpa o currentNumber para o próximo número
-                    _ResultText.value = "0.0" // Prepara para o próximo número
+                    currentNumber = null
+                    _ResultText.value = "0.0"
                 }
-                Log.i("dividir antes / atual", previousNumber.toString() + " / " + currentNumber.toString())
-                return
             }
 
             "√" -> {
@@ -153,21 +160,21 @@ class Functions : ViewModel() {
                     }
                 }
 
-                Log.i("raiz antes / atual", previousNumber.toString() + " / " + currentNumber.toString())
-                return
             }
 
             "CE" -> {
                 if (isCalculatorOn) {
 
-                    _ResultText.value = _ResultText.value?.dropLast(1)?.ifEmpty { "0.0" }
-
-                    if(_ResultText.value?.endsWith(".") == true){
+                    if(_ResultText.value?.endsWith(".") == true && _ResultText.value?.startsWith("0") == true){
                         _ResultText.value += "0"
+                    } else {
+                        _ResultText.value = _ResultText.value?.dropLast(1)?.ifEmpty { "0.0" }
+                    }
+
+                    if (_ResultText.value?.endsWith(".") == true){
+                        _ResultText.value = _ResultText.value?.dropLast(1)?.ifEmpty { "0.0" }
                     }
                 }
-                Log.i("CE antes / atual", previousNumber.toString() + " / " + currentNumber.toString())
-                return
             }
 
             "+/-" -> {
@@ -181,8 +188,6 @@ class Functions : ViewModel() {
                         _ResultText.value = currentNumber.toString()
                     }
                 }
-                Log.i("+/- antes / atual", previousNumber.toString() + " / " + currentNumber.toString())
-                return
             }
 
             "%" -> {
@@ -194,8 +199,6 @@ class Functions : ViewModel() {
                         _ResultText.value = currentNumber.toString()
                     }
                 }
-                Log.i("% antes / atual", previousNumber.toString() + " / " + currentNumber.toString())
-                return
             }
 
             "=" -> {
@@ -205,16 +208,14 @@ class Functions : ViewModel() {
                         currentNumber = performOperation(previousNumber!!, currentNumber!!, operator!!)
                         previousNumber = _ResultText.value?.toDoubleOrNull()
                         _ResultText.value = currentNumber.toString()
-                        operator = null // Reseta o operador após a operação
+                        operator = null
                     }
                 }
-                Log.i("= antes / atual", previousNumber.toString() + " / " + currentNumber.toString())
-                return
             }
 
             "•" -> {
                 if (isCalculatorOn) {
-                    // Adiciona um ponto decimal se não existir um
+                    // Adiciona . se não existir um .
                     if (!_ResultText.value?.contains(".")!!) {
                         // Se o resultado atual for "0.0", substitui por "0."
                         if (_ResultText.value == "0.0") {
@@ -224,20 +225,20 @@ class Functions : ViewModel() {
                         }
                     }
                 }
-                Log.i("decimal antes / atual", previousNumber.toString() + " / " + currentNumber.toString())
-                return
             }
         }
 
         if (isCalculatorOn) {
-            // Concatenar números
-            if (_ResultText.value == "0.0") {
-                _ResultText.value = btn
-            } else {
-                _ResultText.value += btn
+            // Verificar se o botão pressionado é um número ou um ponto decimal antes de atualizar o display
+            if (btn.matches(Regex("[0-9•]"))) {
+                if (_ResultText.value == "0.0") {
+                    _ResultText.value = btn
+                } else {
+                    _ResultText.value += btn
+                }
             }
-            Log.i("Number -> ", previousNumber.toString() + " / " + currentNumber.toString())
         }
+
     }
 
     private fun updateLastTwoButtons(btn: String) {
@@ -246,7 +247,6 @@ class Functions : ViewModel() {
             lastTwoButtons.removeAt(0)
         }
         lastTwoButtons.add(btn)
-        Log.i("Last two Buttons", "$lastTwoButtons -> $btn")
     }
 
     private fun performOperation(num1: Double, num2: Double, op: String): Double {
@@ -255,7 +255,6 @@ class Functions : ViewModel() {
             "-" -> num1 - num2
             "x" -> num1 * num2
             "÷" -> if (num2 == 0.0) {
-                // Tratar a divisão por zero
                 throw ArithmeticException("Cannot divide by zero")
             } else {
                 num1 / num2
